A whole bunch of updates
This can now query for areas and generate stats about where the artists
are from that I'm listening to.

It's not the most intuitive to do and requires a bunch of command line
args and other crap, so maybe at some point I'll come back and redo all
that.

But this is now just a POC that I can bring into the next thing I do.